separate get_tasks from main to mod display_tasks
